Backend/requirements.txt
fastapi
uvicorn
torch
torchvision
ultralytics
pillow
numpy
requests
python-multipart